+++ SOD.Core/Sensor/SensorExtensions.cs
@@ -6,7 +6,7 @@ namespace SOD.Core.Sensor
 {
     static class SensorExtensions
     {
-        public static double CodeToValue(this ICodeBasedSensor sensor, double minValue, double maxValue, int minCode, int maxCode, int code)
+        public static double CodeToValue(this ICodeBasedSensor sensor, double minValue, double maxValue, double minCode, double maxCode, double code)
         {
             if (maxCode == minCode) return 0;
             var k = (maxValue - minValue) / (maxCode - minCode);
